Frd 253 authentification UI

+++ frontend/src/app/core/auth/auth.store.ts
@@ -7,13 +7,17 @@ import {
     withState,
 } from '@ngrx/signals';
 import { UserAuth } from '../models/user.models';
-import { computed, inject, signal } from '@angular/core';
+import { computed, inject, PLATFORM_ID } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
 import { CookieService } from 'ngx-cookie-service';
 import { environment } from '../../../environments/environment.development';
 import { toast } from 'ngx-sonner';
 import { firstValueFrom } from 'rxjs';
 import { AuthService } from './auth-service';
 import { Router } from '@angular/router';
+import { RoleEnum } from '../enums/role.enum';
+import { LoginRequest, LoginResponse } from './auth.models';
+import { ApiResponse } from '../models/response.models';
 
 export interface AuthState {
     user: UserAuth | null;
@@ -33,27 +37,71 @@ export const AuthStore = signalStore(
     withState(initialState),
 
     withComputed(({ token, user }) => ({
-        isAuthenticated: signal(true),
-        userRole: computed(() => user()?.role || 'TOURIST'),
+        isAuthenticated: computed(() => !!token()),
+        userRole: computed(() => user()?.role || RoleEnum.VISITOR),
         userPhoto: computed(() => user()?.photo || 'assets/defaults/default-profile.png'),
     })),
 
     withMethods(
         (
             store,
+            platformId = inject(PLATFORM_ID),
             cookieService = inject(CookieService),
             authService = inject(AuthService),
             router = inject(Router),
         ) => ({
-            setLogin(token: string, user: UserAuth) {
+            async login(credentials: LoginRequest): Promise<boolean> {
+                if (!isPlatformBrowser(platformId)) return false;
+
                 patchState(store, { loading: true });
 
-                const expires = new Date();
-                expires.setMinutes(expires.getMinutes() + 15);
+                try {
+                    const response: ApiResponse<LoginResponse> = await firstValueFrom(
+                        authService.login(credentials),
+                    );
+
+                    if (response.status === 200 && response.data) {
+                        const userData: UserAuth = {
+                            uuid: response.data.uuid,
+                            username: response.data.username,
+                            fullName: response.data.fullName,
+                            photo: response.data.photo,
+                            role: response.data.role,
+                        };
+
+                        this.setLogin(response.data.accessToken, userData);
+
+                        toast.success('Welcome back to DesertAkal!');
+
+                        patchState(store, { loading: false });
+
+                        await router.navigate(['/']);
+
+                        return true;
+                    }
+
+                    patchState(store, { loading: false });
+                    return false;
+                } catch (error: any) {
+                    const msg =
+                        error?.error?.message || 'Login failed. Please check your credentials.';
+                    toast.error(msg);
+
+                    patchState(store, { loading: false });
+
+                    return false;
+                }
+            },
+
+            setLogin(token: string, user: UserAuth) {
+                if (!isPlatformBrowser(platformId)) return;
 
                 const isSecure = environment.secureCookie;
 
-                cookieService.set('auth_token', token, expires, '/', '', isSecure, 'Strict');
+                const tokenExpires = new Date();
+                tokenExpires.setMinutes(tokenExpires.getMinutes() + 15);
+                cookieService.set('auth_token', token, tokenExpires, '/', '', isSecure, 'Strict');
+
                 cookieService.set(
                     'user_data',
                     JSON.stringify(user),
@@ -64,16 +112,13 @@ export const AuthStore = signalStore(
                     'Strict',
                 );
 
-                patchState(store, {
-                    token,
-                    user,
-                    loading: false,
-                });
+                patchState(store, { token, user });
             },
 
             setRefreshToken(newToken: string) {
-                const isSecure = environment.secureCookie;
+                if (!isPlatformBrowser(platformId)) return;
 
+                const isSecure = environment.secureCookie;
                 const expires = new Date();
                 expires.setMinutes(expires.getMinutes() + 15);
 
@@ -81,31 +126,52 @@ export const AuthStore = signalStore(
 
                 const currentUser = store.user();
                 if (currentUser) {
-                    cookieService.set(
-                        'user_data',
-                        JSON.stringify(currentUser),
-                        30,
-                        '/',
-                        '',
-                        isSecure,
-                        'Strict',
-                    );
+                    cookieService.set('user_data', JSON.stringify(currentUser), 30, '/', '', isSecure, 'Strict',);
                 }
 
                 patchState(store, { token: newToken });
             },
 
+            handleOAuthSuccess(data: any) {
+                patchState(store, { loading: true });
+
+                const isSecure = environment.secureCookie;
+                const expires = new Date();
+                expires.setMinutes(expires.getMinutes() + 15);
+
+                cookieService.set('auth_token', data.token, expires, '/', '', isSecure, 'Strict');
+                const user: UserAuth = {
+                    uuid: data.userUuid,
+                    username: data.username,
+                    fullName: data.fullName,
+                    role: data.role,
+                    photo: data.photo
+                };
+                
+                cookieService.set('user_data', JSON.stringify(user), 30, '/', '', isSecure, 'Strict',);
+
+                patchState(store, { 
+                    user: user, 
+                    token: data.token, 
+                    loading: false
+                });
+            },
+
             async logout() {
-                this.setLoading(true);
+                if (!isPlatformBrowser(platformId)) return;
+
+                patchState(store, { loading: true });
+
                 try {
-                    await firstValueFrom(authService.logout());
-                } catch (error) {
-                    toast.error('Server logout failed, cleaning local storage anyway.');
-                } finally {
+                    const response = await firstValueFrom(authService.logout());
+                    toast.success(response.message);    
+
                     cookieService.delete('auth_token', '/');
                     cookieService.delete('user_data', '/');
                     patchState(store, initialState);
-                    router.navigate(['/login']);
+                    await router.navigate(['/auth/login']);
+                } catch (error) {
+                    toast.error('Server logout failed, cleaning local storage anyway.');
                 }
             },
 
@@ -116,17 +182,20 @@ export const AuthStore = signalStore(
     ),
 
     withHooks({
-        onInit(store, cookieService = inject(CookieService)) {
+        onInit(store, platformId = inject(PLATFORM_ID), cookieService = inject(CookieService)) {
+            if (!isPlatformBrowser(platformId)) return;
+
             const savedToken = cookieService.get('auth_token');
             const savedUserJson = cookieService.get('user_data');
 
             if (savedToken && savedUserJson) {
                 try {
                     const user = JSON.parse(savedUserJson) as UserAuth;
-
                     patchState(store, { token: savedToken, user });
-                } catch (error) {
-                    store.logout();
+                } catch {
+                    cookieService.delete('auth_token', '/');
+                    cookieService.delete('user_data', '/');
+                    patchState(store, initialState);
                 }
             }
         },
